SCA: remove padding during parsing
In order to correctly decode Address-Values subfield even number of
semi-octets should be extracted (we should extract full octets only).
But in the same time this field could contains padding semi-octet 'F' if
the actual number contains the odd number of digist.

To avoid saving the padding semi-octet (and returning to user) we should
preserve Address-Length value, then extract rounded up number of
semi-octets and then take only Address-Length number of digits after the
semi-octets swapping.

SCA field is a special case since it contains not a digits number but
an octets number. So add padding semi-octet detection and reduce the
address length by one if we found the padding semi-octet.