SCA: fix isAddress attribute lost while parsing
isAddress attribute is not passed by the PDU address parser at the
moment of SCA object phone setting. So SCA is unable to correctly encode
SMSC address back to the PDU field: it prefixes address data with a
number of digits instead of octets number.

E.g. if you parse a PDU which begins with 07911326040000F0... then if
you try to convert resulted object back to PDU then you got:
0B911326040000F0

So just pass the isAddress parser() argument to the setPhone() method in
order to fix this issue. Actually isAddress should be inverted before
passing to the setPhone() since the method expects a flag which
indicates that the address is for SCA field.