Part: fix the parsed header loss
The Part class constructor interpret passed header argument only as a
Header class constructor params, so it attempts to use header object
passed from the parser as a params to create a new header object. This
behaviour leads to UDH contents loss (e.g. message position in a series
of concatenated messages is not preserved).

To fix this issue check the type of the header argument and if the
argument is a Header object then use it as is. And add tests for header
parsing verification.