tests: add expected error verification support
Add an ability to check error message of exception against the expected
error message, which should be specified in the test case. This option
could be helpful to test cases which rise an expected exceptions (e.g.
this feature will be used in the upcoming tests of the message parts
appending functionality).