Data: append: checks number of parts too
Despite it non-uniformly distribution, the field indicating the number
of parts of the concatenated message, can be used to check whether new
part belongs to the same message as known parts or not.

Therefore, improve the message reference verification with the parts
number check, to make the merging procedure resistant to collision of
reference numbers. Add test for this case as well.