Part: parse: rework input data length calculation
Rework calculation of length of input data (hex string) that should be
processed with an appropriate decoder:
  * explicitly convert septets to octets in case of 7 bit encoding;
  * keep the length in octets and multiply it by 2 at the moment of hex
    stream extraction;
  * account the size of the User Data Header;
  * etc.

Current code attempts to extract more octets than the actual size of
User Data. But since the User Data is the final part of the stream and
the stream splitter returns not more than requested ammount of octets,
the oversized read do not trigger any errors. So, this change do not
fix any errors and just makes the length calculation correct and clear.