Add length argument to the 7-bit decoder
7-bit alphabet decoder needs an input length (in septets) in order to
decode it correctly. Heuristic detection of the length by the decoder
itself gives a good enough result, but it has some issue.

The encoded text must be padded up to the octet boundary. If the text
length modulo 8 is not equal to 7, then the padding takes less than 7
bits and can not cause any issues when decoding. But if the text length
modulo 8 is 7, then the padding takes one full septet. Each code in the
default GSM 7-bit alphabet represents some printable character, and
there are no codes with a special meaning of End-of-String, like a zero
byte in the ASCIIZ-formated strings. Even the code 0000000b in the GSM
alphabet is coresponds to the '@' symbol.

Without an a priori knowledge of the input length (in septets), each
time the input stream length in octets is a multiply of 8 (in other
words, if the padding takes a full septet), the decoder encounters
ambiguity. In such cases, the decoder should guess wheter the final
septet is the padding or a meaning symbol.

If the decoder decides to ignore the final septet, it will lose the last
symbol of the text. If the decoder assumes that the final septet is a
symbol, then in many cases the decoded text is wrongly appended with the
'@' symbol (since the use of zero bits for padding is mandatory in
accordance with 3GPP TS 23.038). If the decoder ignores the final
septet, only if it is zero (our case), then it produces the correct
output for most of the observed messages. But sometimes the SM text is
really ends with '@'.

To avoid this set of issues and turn the parsing to bulletproof, caller
code should derive the encoded text length (in septets) from an
available length information (e.g. use the UDL in the UD parser, use the
Address-Length in the Address parser, etc.) and pass the length to the
decoder. On the decoder side, if the length is passed, use it, otherwise
the decoder will fall back to the old model, checking the value of the
final septet.

Also add a few tests for the cases, which are specified above.