Fix extended 7-bit character decoding in the final position
The decoder does not check whether the code should be processed using
the extended table when processing the final septet. This will lead to
incorrect decoding of the symbol from the extended table, if it appear
in the final position (e.g. '}' will be decoded as ')'). To fix this,
perform regular decoding (using the extended table) of the final septet
too.