Rework the 7-bit alphabet decoder
Rework the decoder of the 7-bit alphabet to make it clean and reduce
duplication of code within the decoder's main loop (3 places of the
septet processing are reduced to one). The main difference from the
previous version is the explicit use of buffer in a FIFO manner instead
of workarounds with carry bits.