Header: stop random message reference generation
This change modify the default behaviour of the Header class
constructor. If no initialization parameters are provided then
constructor no more generates the message reference number (for messages
concatenation) by itself.

This change is required for a further class rework to add a support of
User Data Headers of arbitrary contents.

Rationalization for the default behaviour change is simple: message
reference number should be the same for each part of concatenated
message, so only code of upper levels should generate this number and
set it for each part  (e.g. Data class is doing in such way). So we just
do not need number generation functionality in the Header class.