Header: allow TYPE be undefined and dynamic size
Make TYPE undefined by default and set it to 16 bit ref. num type if the
caller provides parameters at the moment of object creation. This allows
us to clearly indicate that there are no concatenation data was
configured.

Also drop the internal fields for the UDHL and the pointer size tracking
and generates these values dynamically in dependance of TYPE
configuration state.

And finally, generate an empty header if the TYPE was not configured.