Header: add Information Elements array
User Data Header (UDH) of PDU is consist of separate Information
Element(s) (IEs) and UDH may includes arbitraty number of IE(s) of
various types since the IE(s) are formated as a typical TLV. To fully
support UDH parsing we need a some structure, which allows us to keep
arbitrary IE(s) and work with them.

So add an internal field of the array type to Header class to solve
this. Each array element represents one IE in form of object with at
least two fields:
 * type - the IE type
 * dataHex - the IE data in a form of raw hexadecimal string

This is a preparation step. Later changes will make this array a main
element of the whole class.