Header: place concat data within IE
Move fields that keep cache of concatenation data (e.g. number of parts,
current part position, etc.) to the coresponding IEs array element. And
use concat IE index to fetch this data when we need it.

This change should reduce the ammount of specific fields in the Header
class and make data storage less puzzled.