Make 7-bit decoder aware of align bits
If the encoding method is the 7-bit default GSM alphabet then the text
after the User Data Header should be aligned to the septet boundary.

It is just a coincidence that the overall size (UDHL + UDH) of UDH
with 16bit ref. number is 7 octets and so the UDH end is equal to septet
boundary. In other cases attempting to decode text immediatly after the
UDH leads to an invalid decoding result.

To fix this calculate the alignment length in the Part parser and pass
this value to the 7-bit alphabet decoder. On the decoder side skip
the alignment bits before begin the real text decoding. Add a new test
for this case.