Make 7-bit encoder aware of align bits
If the encoding method is the 7-bit default GSM alphabet then the text
after the User Data Header should be aligned to the septet boundary.

To facilitate alignment insertion add a new argument to 7-bit encoder
and insert the requested ammount of zero bits to the produced encoder
output.

This change is mainly required to support testing of decoding text with
leading alignment (which is included in this change too). Because the
test code perfoms both encoding and decoding operations for each test
case and we siply can not test the decoder if we unable to run encoder
with the same parameters.