SCA: parser: make variable usage consistent
The variable named 'octets' is actually used to keep the number of
semi-octets (which is equal to the number of hexadecimal digits in the
input string). This can be misguiding for the user who read the code the
first time.

The calculations was reworked in such way that the variable 'octets' now
contains the length of the address value in octets and the
multiplication x2 (conversion to the number of hex chars) is performed
at the moment of extraction of the hexadecimal string. Also add a couple
of comments, to make it clear why the commented transformation is being
performed.

This change does not fix any issues. It is a kind of cosmetic change
that should make the code more clear.

Fixes: 3606010 ("SCA: remove padding during parsing")