SCA: fix encoding SCA of even length
The method used to calculate the division with the ceiling ((x + 1) / 2)
is good enough if the division is an integer operation. But JS division
produces a floating point result, and such calculation corrupts the SCA
field generation. To fix this, explicitly call Math.ceil() method.

Before this change, the library encodes SCA address '999999999999' (from
the readme) as:
  7.891999999999999
after the change it produces:
  0791999999999999