DCS: disable compression by-default
The library does not yet support compression algorithms from 3GPP TS
23.042: neither compression or decompression. But compression is
activated by default and even implicitly activated if the user sets the
alphabet to the 7-bit default.

Such misguiding behaviour of the library did not hurt anyone by a simple
cause: the encoding of the DCS compressed flag was broken (this will be
fixed in the next commit).

In order to prevent misguiding output generation, disable compression by
default.