DCS: fix compressed and class ind. flags handling
During the DCS value construction (as well as during parsing), the DCS
octet is splitted onto the couple of nibbles (aka semi-octets, aka
subfields): encodeGroup (high nibble) and dataEncoding (low nibble). So,
bits 4 and 5 of the DCS become bits 1 and 0 of the encodeGroup. But the
code still tries to set (as well as get) the flags of the text
compression and the class indication at offsets of 5 and 4,
respectively.

Fix the flags handling by using the correct offsets. Add test with a
flash SMS for this case.