tests: attempt to recreate PDU
After a successful parse, try to recreate the PDU (encode it back to a
hexadecimal string) and compare the resulting PDU and the input string.
If the strings are mismatch, then assume that the test fails.

Such test does not give a precise information about error source, but
increases the code  coverage.